+++ trackclient/src/main/java/com/heliossoftwaredeveloper/trackclient/repository/TrackRepository.kt
@@ -3,9 +3,14 @@ package com.heliossoftwaredeveloper.trackclient.repository
 
 import com.heliossoftwaredeveloper.applicationdatabase.dao.TrackDao
 import com.heliossoftwaredeveloper.applicationdatabase.entity.TrackEntity
+import com.heliossoftwaredeveloper.common.util.safeDispose
 import com.heliossoftwaredeveloper.trackclient.TrackApiClient
 import com.heliossoftwaredeveloper.trackclient.model.SearchTrackResponse
+import com.heliossoftwaredeveloper.trackclient.serviceModelToTrackEntity
 import io.reactivex.Observable
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.Disposable
+import io.reactivex.schedulers.Schedulers
 
 /**
  * Repository class for Track related stuff
@@ -37,23 +42,36 @@ interface TrackRepository {
      *
      * @param trackList the list of track entity to save in database.
      */
-    fun saveSearchTrack(trackList: List<TrackEntity>): Observable<Unit>
+    fun saveSearchTrack(trackList: List<TrackEntity>?)
 }
 
 class TrackRepositoryImpl(private val apiClient: TrackApiClient, private val trackDao: TrackDao): TrackRepository {
 
+    private var saveTrackDisposable : Disposable? = null
+
     override fun searchTrack(term: String, country: String, media: String): Observable<SearchTrackResponse> {
         return apiClient.getService().searchTrack(term, country, media)
+            .subscribeOn(Schedulers.io())
+            .observeOn(AndroidSchedulers.mainThread())
+            .doOnNext { saveSearchTrack(it?.results?.serviceModelToTrackEntity()) }
     }
 
     override fun getLastSearchTrack(): Observable<List<TrackEntity>> {
         return trackDao.getAllTracks()
+            .subscribeOn(Schedulers.io())
+            .observeOn(AndroidSchedulers.mainThread())
     }
 
-    override fun saveSearchTrack(trackList: List<TrackEntity>): Observable<Unit> {
-        return Observable.fromCallable {
+    override fun saveSearchTrack(trackList: List<TrackEntity>?){
+        saveTrackDisposable = Observable.fromCallable {
             trackDao.deleteTracks()
-            trackDao.saveTracks(trackList)
+            trackList?.let {
+                trackDao.saveTracks(it)
+            }
         }
+            .subscribeOn(Schedulers.io())
+            .observeOn(AndroidSchedulers.mainThread()).subscribe {
+                saveTrackDisposable?.safeDispose()
+            }
     }
 }
