+++ trackui/src/main/java/com/heliossoftwaredeveloper/trackui/util/TrackModelMapper.kt
@@ -82,40 +82,3 @@ fun List<TrackDetailResult>.serviceModelToTrackItem() : List<TrackItem> {
         )
     }
 }
-
-/**
- * Extension function to transform track api service model to entity database model
- *
- * @return List<TrackEntity> The transformed TrackEntity model
- */
-fun List<TrackDetailResult>.serviceModelToTrackEntity() : List<TrackEntity> {
-    return this.map {
-        TrackEntity(
-            id = 0,
-            kind = it.kind,
-            trackId = it.trackId,
-            artistName = it.artistName,
-            collectionName = it.collectionName,
-            trackName = it.trackName,
-            artworkUrl = it.artworkUrl100,
-            collectionPrice = it.collectionPrice,
-            trackPrice = it.trackPrice,
-            trackRentalPrice = it.trackRentalPrice,
-            collectionHdPrice = it.collectionHdPrice,
-            trackHdPrice = it.trackHdPrice,
-            trackHdRentalPrice = it.trackHdRentalPrice,
-            releaseDate = it.releaseDate,
-            collectionExplicitness = it.collectionExplicitness,
-            trackExplicitness = it.trackExplicitness,
-            trackCount = it.trackCount,
-            trackNumber = it.trackNumber,
-            trackTimeMillis = it.trackTimeMillis,
-            country = it.country,
-            currency = it.currency,
-            primaryGenreName = it.primaryGenreName,
-            contentAdvisoryRating = it.contentAdvisoryRating,
-            shortDescription = it.shortDescription,
-            longDescription = it.longDescription
-        )
-    }
-}
